Refactored file names; added provider wraps to ApplicationViews for testing
