+++ src/components/friends/FriendForm.js
@@ -1,0 +1,1 @@
+
