+++ src/components/friends/FriendItem.js
@@ -1,0 +1,1 @@
+
