+++ src/components/friends/FriendList.js
@@ -1,0 +1,1 @@
+
