--- src/components/friends/FriendsForm.js
@@ -1,1 +1,0 @@
-
