--- src/components/friends/FriendsItem.js
@@ -1,1 +1,0 @@
-
