--- src/components/friends/FriendsList.js
@@ -1,1 +1,0 @@
-
