+++ src/components/messages/MessageProvider.js
@@ -1,0 +1,39 @@
+import React, { useState, createContext } from "react"
+
+export const MessageContext = createContext()
+
+export const MessageProvider = (props) => {
+    const [messages, setMessages ] = useState([])
+
+    const getMessages = () => {
+        return fetch("localhost:8088/messages")
+        .then(response => response.json())
+        .then(setMessages)
+    }
+
+    const addMessage = messageObj => {
+        return fetch("localhost:8088/messages", {
+            method: "POST",
+            headers: {
+                "Content-Type:": "application.json"
+            },
+            body: JSON.Stringify(messageObj)
+        })
+        .then(getMessages)
+    }
+
+    const removeMessage = id => {
+        return fetch(`localhost:8088/messages/${id}`, {
+            method: "DELETE"
+        })
+        .then(getMessages)
+    }
+
+    return (
+        <MessageContext.Provider value={{
+            messages, getMessages, addMessage, removeMessage
+        }}>
+            {props.children}
+        </MessageContext.Provider>
+    )
+}
