--- src/components/messages/MessagesForm.js
@@ -1,1 +1,0 @@
-
