--- src/components/messages/MessagesItem.js
@@ -1,1 +1,0 @@
-
