--- src/components/messages/MessagesList.js
@@ -1,1 +1,0 @@
-
