+++ src/components/plannedEvents/PlannedEventProvider.js
@@ -2,8 +2,8 @@ import React, { useState, createContext } from "react"
 
 export const PlannedEventContext = createContext()
 
-export const PlannedEventProvider = () => {
-    const [plannedEvents, getPlannedEvents ] = useState([])
+export const PlannedEventProvider = (props) => {
+    const [plannedEvents, setPlannedEvents ] = useState([])
 
     const getPlannedEvents = () => {
         return fetch("localhost:8088/plannedEvents")
