--- src/components/tasks/TasksList.js
@@ -1,1 +1,0 @@
-
