+++ src/components/users/UserProvider.js
@@ -1,0 +1,25 @@
+import React, { useState, createContext } from "react"
+
+export const UserContext = createContext()
+
+
+export const UserProvider = (props) => {
+    const [ users, setUsers ] = useState([])
+
+    const getUsers = () => {
+        return fetch("localhost:8088/users")
+        .then(response => response.json)
+        .then(setUsers)
+    }
+
+    return (
+        <UserContext.Provider value={{
+            users, getUsers
+        }}>
+            {props.children}
+        </UserContext.Provider>
+    )
+
+
+    
+}
